Build a basic automcomplete component
I used this blog post https://alligator.io/react/react-autocomplete/ for
some inspiration and followed it for the most part (adjusting a bit for
preact and typescript and my own style). Drag and drop tags look like
they may be a bit more involved though, so I'm going to stash this on a
branch and try installing a few react libraries instead to see if that
works just as well.