Implement a tagselector input with autocomplete and tag lists
Tags are pretty central to how we currently measure food similarity, so
making it easy to edit and curate tags is important and should be easy
for moderators. There are libraries out there that provide prettier
support with drag and drop behavior, but I just wanted to get something
simple operational. The tag selector shouldn't be styled completely with
the assumption that it will only be used in conjunction with adding and
removing tags from a food tag list -- it should also be reuseable for
the food search filter (which is mainly based on filtering for foods
that contain a tag).

For now I've just hardcoded a list of some tags, but ultimately this too
should be coming from a more dynamic and moderatable backend list of
tags.