Add tag detail, start refactoring
I've cleaned up the data model on firebase a bit and added in some
common visualization for where tags and foods fit in a food space --
using a fingerprint barchart concept along with a list of nearest
neighbors. Meanwhile I've cleaned up the tags on firebase so that the
foods are consistently tagged with the cannonical tag representation
(it's ID) which is separate from the pretty version (which ultimately
could turn into a localized language specific version if we choose).

I've started to refactor the typescript type interfaces over to some
common shared types, but I'm hitting a number of issues with my sloppy
handling of (p)react state vs props. I need to clean that up for both
tags and foods. I also think it might be time to create a separate
folder for each.