Fixed firestore overusage and added pagination
The way the tag selector was implemented -- loading one firestore doc
per tag -- was quickly pushing us over the 50k daily document read limit
in our free tier of firebase. I've created a new document in the new
collection called "indexes" which for now simply holds an array of all
of the tag ids. That way when we load the tag selector, we just have to
load one document rather than 1k documents.

The food list was another page that was quickly eating up reads as well
-- every time somebody visits the default page it would load all food
documents (currently only 270). I've added in some pagination to limit
the load to 20 pages per request, and rather than actual pages there's
simply a more button to add an additional 20 foods to the list if there
are more available.