Filter foods using all tag types
Firebase only has support for relatively simplistic query filters. We
made the decision to separate tags into three different groups -- this
may not end up being all that useful in the end, but it has made it more
difficult to filter foods based on tags because we can't simply query
for the tag in ANY of the arrays. As a simple hack to address this, I've
added an extra array of tags that contains the union of all of the other
arrays. This will make submitting updates a little more complicated
since we'll have to automatically rebuild the union, but at least it
will let our filter do what we need it to do.