Add some welcome text to the home page
I'm not sure it even makes sense to have one, maybe we just want to
redirect straight into the foods tab. For now I'm leaving it there for
some instructions and a basic place holder.

I also added a header to the foods page to bring some attention to the
tag filter input box that was unlabled up until now.