Add sortable tagstats to tag list
The tag list page was just a long list of all of our tags. In order to
get a better idea of which tags were used a lot and which were likely
erroneous or not valuable, we'd like to see frequency counts (how often
they're used on foods). It would also be very helpful if these were
sortable columns in a table.

While I was at it, I figured it would also be nice if we could sort tags
by the semantic dimensions that we've extracted and display in the
fingerprint for foods an images. I made an attempt to name the
dimensions with semi interpretable sides, then I added those to the
fingerprint component to make it easier to interpret.