Persist tag sort on back nav
The tags list page supports multiple sort options, but if you sorted the
page and clicked on one of the links, then went back to the page it
would resume the default sort. This was moderately annoying if you were
interested in looking at a few of the tags at the top of one of the
sorts. I've added the sort details into the route itself and reroute on
sorting so that navigation with the back button will take you back to
your last sort. This will also let you share a particular sort with
another user as part of the url. I decided not to save each sort as a
separate page in the history though -- I don't expect people want to
have to click back through every one of their sorts to get further back
in their history.