Support filtering foods by tag types
Up until now the filtered food page showed all foods that had any tag
that matched the filter (regardless of tag type). It would be nice if we
could specify the tag type that we want to look at during the filter,
but I didn't want to complicate the UI too much. I realized that the
tag stats page provided a bit of a solution to the UI complexity though
-- we can leave the basic filter as is, but treat each of the tag
frequencies in the tag stats page as a link to the filtered food page
with the tag type already specified. No additional fixtures, but finer
grain filtering is still available if you go by way of the stats page.