Make food list route sortable by dimensions
Also port the implementation over to one similar to the tag list route
using a single firestore document for the stats. In this case we're
using two single document indexes -- one for the sortable stats and one
for the inverted index. This lets us sort and filter without having to
query individual foods from the food collection. I'm abandoning the
ability to filter by tag type for now.

Also, this was a quick hack -- it's ugly. I'll have to clean things up
later.