Track last time food updates were made and make sortable
We not only need to update the food document itself, but also the
foodspace that we use for sorting foods on the food list route. This
will make it much easier for moderators to find the foods that they
haven't yet currated (or in the future ones that have recently been
added).

Separate from these changes it took a while to figure out how to deal
with importing and exporting firestore timestamp objects. Internally
they're stored as a custom object that's composed of two integers
(seconds since the unix epoch and nano seconds from that point) but
using the default JSON stringify and parse methods to serialize and
deserialize that object render it unusable. I've addressed that in the
ladle projects import/export methods by converting it into an ISO
timestamp string in the export and then converting any strings with the
ISO timestamp format back into firestore timestamps on import.