Port tag selector over to tagFoods index
The tagIds firestore list of tags was an early implementation that's now
being phased out as we move over to using the inverted index and tag
stats index more throughout the app. It's nice having an especially
small list of tags, but it gives us one more index to try to have to
maintain.