Gracefully handle addition of new tags
For now the tag selector lets you enter a tag that isn't in the drop
down list. If you submit the food form with that new tag, we want to
make sure that we don't break the tag list route (sane stats/counts
should show up for that new tag and they should be sortable) or the
route for that tag in general.

We should at the very least add in some validation to make sure wonkey
characters don't end up in the tag names.