Track automatic food updates separately from manual edits
When I added in the option to delete tags and automatically have them
removed from any foods that they appeared in, I forgot to update the
last changed timestamp on those foods. That would be pretty critical for
incremental backups -- however, when I went to add that in, I realized
that it would break the ability to sort foods by when they were last
edited -- something that we'd like among other reasons for making it
easier to find foods whose metadata we haven't yet fixed. Those are
exactly the foods most likely to have bogus tags that we might use the
bulk tag deletion option on, so using the option would make it harder to
find them for additional manual edits (like fixing the title).