Add support for highlighting secondary tag
We'd like to coopt tags for non appetite related purposes -- like
labeling which foods to include in an mvp for instance. In order to
support the curation of diverse baskets of food though, our moderators
will want to be able to see compare foods with similar food related tags
that are and are not already included in that basket. If the basket is
just another tag, that means we want some way of visualizing the
interaction between tags.

Highlighting was the simplest approach we could think of for now --
filter by a tag like Pizza and highlight the basket tag (like _mvp_).
Then compare the available images of pizza that are in the mvp basket to
those that aren't, select the appropriate number and variety and move
on.

This commit is mostly just a proof of concept. It needs to be refined,
cleaned up, probably optimized a bit, and extended beyond just the food
list route (e.g. similar foods, tag route foods).