Refactor indices firestore data
When we tried to include all of the images, we started hitting size
problems on the indexes collection. The problem wasn't exactly a matter
of size, but rather too many fields to "index". These documents are
themselves individual indexes and don't need to be indexed -- we would
never query for documents from the indexes collection where X=Y, so we
can actually disable the indexing that firestore is automatically doing.
Unfortunately, there's a limit of 200 "exemptions" per collection and
you have to manually list each of the fields that you want to disable
indexing for as it's own exemption. I saw a recommendation somewhere to
just stick all of the fields that you want to disable indexing for into
a map that is the value of a single field, then disable that field -- so
that's why we've dumped the entire contents of our indexes into a field
called "data" and disabled indexing for that field.

As part of the process I've also renamed the foodspace index to
foodStats to be more consistent with the tagStats index naming and
changed the collection name to indices from indexes to allow them both
to exist simultaneously while we port over.